#45 Break pmove out to game/cgame. This prompted me to refactor all of the client-side collision code such as R_Trace and R_PointContents, which are gone. Cl_PointContents and Cl_Trace now provide more complete and correct implementations by factoring in bounded-box entities and supporting the 'skip' entity. I also noticed along the way that Sv_Trace used a different ordering of parameters from Cm_Trace, which is retarded. That rippled through to the game code, which was unfortunate. So that's all fixed now too.

+++ src/game/default/g_client_view.c
@@ -70,8 +70,8 @@ static void G_ClientWaterInteraction(g_edict_t *ent) {
 		return;
 	}
 
-	const int32_t water_level = ent->locals.water_level;
-	const int32_t old_water_level = ent->locals.old_water_level;
+	const uint8_t water_level = ent->locals.water_level;
+	const uint8_t old_water_level = ent->locals.old_water_level;
 
 	ent->locals.old_water_level = water_level;
 
@@ -84,8 +84,7 @@ static void G_ClientWaterInteraction(g_edict_t *ent) {
 		gi.Sound(ent, gi.SoundIndex("world/water_out"), ATTN_NORM);
 
 	// head just coming out of water, play a gasp if we were down for a while
-	if (old_water_level == 3 && water_level != 3 && (client->locals.drown_time - g_level.time)
-			< 8000) {
+	if (old_water_level == 3 && water_level != 3 && (client->locals.drown_time - g_level.time) < 8000) {
 		vec3_t org;
 
 		VectorAdd(client->ps.pm_state.origin, client->ps.pm_state.view_offset, org);
@@ -118,8 +117,8 @@ static void G_ClientWaterInteraction(g_edict_t *ent) {
 			client->locals.pain_time = g_level.time;
 
 			// and apply the damage
-			G_Damage(ent, NULL, NULL, vec3_origin, ent->s.origin, vec3_origin, ent->locals.dmg, 0,
-					DAMAGE_NO_ARMOR, MOD_WATER);
+			G_Damage(ent, NULL, NULL, vec3_origin, ent->s.origin, vec3_origin, ent->locals.dmg, 0, DAMAGE_NO_ARMOR,
+					MOD_WATER);
 		}
 	}
 
@@ -130,13 +129,13 @@ static void G_ClientWaterInteraction(g_edict_t *ent) {
 			client->locals.sizzle_time = g_level.time + 200;
 
 			if (ent->locals.water_type & CONTENTS_LAVA) {
-				G_Damage(ent, NULL, NULL, vec3_origin, ent->s.origin, vec3_origin, 2 * water_level,
-						0, DAMAGE_NO_ARMOR, MOD_LAVA);
+				G_Damage(ent, NULL, NULL, vec3_origin, ent->s.origin, vec3_origin, 2 * water_level, 0, DAMAGE_NO_ARMOR,
+						MOD_LAVA);
 			}
 
 			if (ent->locals.water_type & CONTENTS_SLIME) {
-				G_Damage(ent, NULL, NULL, vec3_origin, ent->s.origin, vec3_origin, 1 * water_level,
-						0, DAMAGE_NO_ARMOR, MOD_SLIME);
+				G_Damage(ent, NULL, NULL, vec3_origin, ent->s.origin, vec3_origin, 1 * water_level, 0, DAMAGE_NO_ARMOR,
+						MOD_SLIME);
 			}
 		}
 	}
